% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grur_dapc_pipeline.R
\name{dapc_clustering}
\alias{dapc_clustering}
\title{grur DAPC and likelihood-based genetic clustering pipeline based on adegenet}
\usage{
dapc_clustering(x, strata = "POP_ID", pop.levels = NULL, n.pca.max = 300,
  n.rep = 100, training.set = 0.9, pc = 1:4, plot.filename = NULL,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{x}{Genind object}

\item{strata}{(character, optional) Default: \code{strata = "POP_ID"}.}

\item{pop.levels}{(character string, optional) Default: \code{pop.levels = NULL}.}

\item{n.pca.max}{(integer, optional) Default: \code{n.pca.max = 300}.}

\item{n.rep}{(integer, optional) Default: \code{n.rep = 100}.}

\item{training.set}{(double, optional) Default: \code{training.set = 0.9}.}

\item{pc}{(integer string, optional) Default: \code{pc = 1:4}.}

\item{plot.filename}{(character, optional) Default: \code{plot.filename = NULL}.}

\item{parallel.core}{(integer, optional) The number of core used for parallel
execution during the pipeline.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
coming
}
\description{
Analyse data with adegenet DAPC and snapclust.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
