% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grur_dapc_pipeline.R
\name{generate_dapc_pc_plot}
\alias{generate_dapc_pc_plot}
\title{generate_dapc_pc_plot}
\usage{
generate_dapc_pc_plot(pc.to.do, pc.data, variance.component,
  plot.filename = NULL)
}
\description{
Generate the PC plots
}
\keyword{internal}
