% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{generate_pcoa_plot}
\alias{generate_pcoa_plot}
\title{generate_pcoa_plot}
\usage{
generate_pcoa_plot(strata.select, pc.to.do, vectors, variance.component,
  path.folder, write.plot)
}
\description{
Generate the PCoA plots
}
\keyword{internal}
