% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequencing_strata.R
\name{generate_sequencing_strata}
\alias{generate_sequencing_strata}
\title{Generate sequencing and lanes strata for testing}
\usage{
generate_sequencing_strata(strata, number.ind.per.lanes = 36,
  number.sequencer = NULL, randomize = FALSE, filename = NULL)
}
\arguments{
\item{strata}{The strata file used in \code{grur}, \code{stackr} and
\code{assigner} is described in \code{\link[stackr]{tidy_genomic_data}}.}

\item{number.ind.per.lanes}{(integer) By giving the number of individuals
you want to pool per lanes,
this will attribute a lane id to individual, sequentially.
Default: \code{number.ind.per.lanes = 36}.}

\item{number.sequencer}{(optional, integer) Give the number of sequencer
used for this strata.
Why sequencer ? so far I've seen numerous projects with missingness pattern
with a sequencing signature. With default, \code{number.sequencer = NULL},
there is no column created in the strata file.}

\item{randomize}{(optional, logical) To randomize the lanes for individuals
and the sequencer used by each lanes.
Default: \code{randomize = FALSE}.}

\item{filename}{(optional) The name of the new strata file
written to the working directory.
Default: \code{filename = NULL}, the strata is only in the global
environment.}
}
\value{
A strata object (dataframe) in the global environment with columns: 
\code{POP_ID}, \code{INDIVIDUALS}, \code{LANES} and optionally \code{SEQUENCER}.
}
\description{
Use this function to generate more stratification inside you're
strata file using sequencer and lanes. This is useful to test hypothesis of
no missingness pattern associated with these variables or to generate missing
genotypes based on these variables.
}
\examples{
\dontrun{
new.strata <- generate_sequencing_strata(
strata = "my.strata.tsv",
number.ind.per.lanes = 48,
number.sequencer = 3,
randomize = TRUE,
filename = "my.new.strata.tsv"
)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
