% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grur_imputations.R
\name{grur_boost_imputer}
\alias{grur_boost_imputer}
\title{grur_boost_imputer}
\usage{
grur_boost_imputer(boost.split = NULL, markers.df = NULL,
  input.xgb = NULL, gl.wide = NULL, eta = 0.2, gamma = 0,
  max_depth = 6, min_child_weight = 1, subsample = 0.8,
  colsample_bytree = 1, num_parallel_tree = 1, cpu.boost = 1,
  nrounds = 200, early_stopping_rounds = 20,
  save_name = "imputation.model.temp")
}
\description{
imputations using eXtreme Gradient Boosting Tree
}
\keyword{internal}
