% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grur_imputations.R
\name{grur_imputer}
\alias{grur_imputer}
\title{grur_imputer}
\usage{
grur_imputer(data, num.tree = 100, pmm = 0, random.seed = NULL,
  parallel.core = parallel::detectCores() - 1,
  hierarchical.levels = "strata", markers.list = markers.list,
  verbose = verbose)
}
\description{
imputations using Ranger package and predictive mean matching
}
\keyword{internal}
