% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputations_accuracy.R
\name{imputations_accuracy}
\alias{imputations_accuracy}
\title{Measure imputation accuracy}
\usage{
imputations_accuracy(obs, imp)
}
\arguments{
\item{obs}{The original dataset prior to imputations. This can be simulated
data.}

\item{imp}{The imputed data with predicted genotyped values.}
}
\value{
A list with information that were dropped from the analysis.
This can be populations, individuals and markers.
The accuracy measure returned is the Root Mean Square Error (RMSE). 
The RMSE is provided by populations, individuals, markers and overall.
}
\description{
Use this function to measure imputations accuracy and track impact
of different algorithms on simulated or biological datasets.
}
\examples{
\dontrun{
rmse.imputations <- imputations_accuracy(
obs = "sim.data",
imp = "imputed.data"
)
}
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
