% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grur_imputations.R
\name{impute_genotypes}
\alias{impute_genotypes}
\title{impute_genotypes}
\usage{
impute_genotypes(markers.list, data, data.na, data.gl = NULL, data.imp,
  num.tree = 100, pmm = 0, random.seed = NULL,
  parallel.core = parallel::detectCores() - 1,
  hierarchical.levels = "strata", pred.error = pred.error)
}
\description{
imputations using Ranger package and predictive mean matching of missRanger
}
\keyword{internal}
