% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{landscape.new.ind.genos}
\alias{landscape.new.ind.genos}
\title{landscape.new.ind.genos}
\usage{
landscape.new.ind.genos(rland, PopulationSizes, AlleleFreqs = NULL)
}
\description{
landscape.new.ind.genos
}
\keyword{internal}
