% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{make_mig_mat}
\alias{make_mig_mat}
\title{make_mig_mat}
\usage{
make_mig_mat(mig.rate, num.pops, type = c("island", "stepping.stone"))
}
\description{
Make migration matrix for simulation
}
\keyword{internal}
