% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_missing.R
\name{markers_dirichlet}
\alias{markers_dirichlet}
\title{markers_dirichlet}
\usage{
markers_dirichlet(number.markers, alpha.beta.markers)
}
\description{
Simulate the Compound Dirichlet-Multinomial (cdm) for each markers
as a bunch of gammas scaled by their sum.
}
\keyword{internal}
