% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_distributions.R
\name{missing_distributions}
\alias{missing_distributions}
\title{Missing data distributions}
\usage{
missing_distributions()
}
\value{
list of parameters
}
\description{
Graphical interface to choose parameters for missing data.
}
\author{
Eric Archer \email{eric.archer@noaa.gov} and 
  Thierry Gosselin \email{thierrygosselin@icloud.com}
}
