% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_missing.R
\name{missing_nmar}
\alias{missing_nmar}
\title{missing_nmar}
\usage{
missing_nmar(tidy, number.individuals, number.markers, average.read.depth,
  min.reads, ind.pop.list, alpha.beta.markers, markers.list)
}
\description{
Function to generate missing NOT at random
}
\keyword{internal}
