% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{missing_rda}
\alias{missing_rda}
\title{missing_rda}
\usage{
missing_rda(
  data,
  strata,
  permutations = 1000,
  parallel.core = parallel::detectCores() - 1
)
}
\description{
The function uses Permutation test Anova and
Redundancy Analysis to highlight potential patterns of missingness in the data
based on the strata provided
}
\keyword{internal}
