% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{pct_missing_by_total}
\alias{pct_missing_by_total}
\title{pct_missing_by_total}
\usage{
pct_missing_by_total(strata.select, data, ci = 0.95, path.folder,
  write.plot = TRUE)
}
\description{
Generates plot missing by total
}
\keyword{internal}
