% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_missing.R
\name{pop_dirichlet}
\alias{pop_dirichlet}
\title{pop_dirichlet}
\usage{
pop_dirichlet(pop.list, number.markers, alpha.markers)
}
\description{
Function to make a different loc.dirichlet for each population
}
\keyword{internal}
