% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rad.R
\name{simulate_rad}
\alias{simulate_rad}
\title{Simulate RADseq data}
\usage{
simulate_rad(num.pops = 5, num.loci = 1000, div.time = 25000, ne = c(50,
  500), nm = c(0, 0.1, 0.5, 1, 5), theta = 0.2, mig.type = c("island",
  "stepping.stone"), num.reps = 10, num.rms.gens = 5, label = NULL,
  fsc.exec = "fsc252", parallel.core = NULL)
}
\arguments{
\item{num.pops}{number of populations.
Default: \code{num.pops = 5}.}

\item{num.loci}{number of independent RADseq loci.
Default: \code{num.loci = 1000}.}

\item{div.time}{time since divergence of populations.
Default: \code{div.time = 25000}.}

\item{ne}{effective populations size.
Default: \code{ne = c(50, 500)}.}

\item{nm}{number of migrants per generation.
Default: \code{nm = c(0, 0.1, 0.5, 1, 5)}.}

\item{theta}{value of theta (= ne * migration rate).
Default: \code{theta = 0.2}.}

\item{mig.type}{migration topology type: \code{"island"} or 
\code{"stepping.stone"}.}

\item{num.reps}{number of replicates for each scenario.
Default: \code{num.reps = 10}.}

\item{num.rms.gens}{number of \code{rmetasim} generations to establish
linkage disequilibrium.
Default: \code{num.rms.gens = 5}.}

\item{label}{label for the output folder. With default \code{label = NULL},
the label will be "\code{sim.results.YYYYMMDD.HHMM}".}

\item{fsc.exec}{name of fastsimcoal executable.
Default: \code{fsc.exec = "fsc252"}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
invisibly, the label used to name output files
}
\description{
Simulate populations of RADseq data following island or stepping 
stone models
}
\note{
The following arguments can be specified as single values or vectors:
 \code{num.pops, num.loci, div.time, ne, nm, theta, mig.type}. One simulation 
 scenario will be generated for each combination of unique values of these 
 arguments.  
 
 The function requires that the executable for \code{fastsimcoal} be installed 
 in a location where it can be executed from the current working directory.  
 
 The function will create a folder in the current working directory with the 
 name provided by \code{label}. Within this folder there will be an R 
 workspace file ("\code{scenarios.rdata}") containing a data frame with the 
 parameters for each scenario (\code{sc.df}), and one R  
 workspace file for each scenario, named \code{"gtypes.#.rdata"} where "#" 
 is the scenario number. Each scenario file contains a list named 
 \code{sim.data}, which contains a \code{strataG::gtypes} representation 
 of the simulated SNP data for every replicate in that scenario. The list 
 has a "\code{scenario}" attribute which is a one row data.frame providing 
 the parameters that generated the data and a "\code{label}" attribute which 
 is the label of the set of scenarios that were run.
}
\author{
Eric Archer <eric.archer@noaa.gov>
}
