% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSNPs.R
\name{simulateSNPs}
\alias{simulateSNPs}
\title{Simulate SNPs}
\usage{
simulateSNPs(num.pops = 5, num.loci = 1000, div.time = 25000, Ne = c(50,
  500), Nm = c(0, 0.1, 0.5, 1, 5), theta = 0.2, mig.type = c("island",
  "stepping.stone"), num.reps = 10, num.rms.gens = 5, label = NULL,
  fsc.exec = "fsc252", num.cores = NULL)
}
\arguments{
\item{num.pops}{number of populations}

\item{num.loci}{number of independent SNP loci}

\item{div.time}{time since divergence of populations}

\item{Ne}{effective populations size}

\item{Nm}{number of migrants per generation}

\item{theta}{value of theta (= Ne * migration rate)}

\item{mig.type}{migration topology type: \code{"island"} or 
\code{"stepping.stone"}}

\item{num.reps}{number of replicates for each scenario}

\item{num.rms.gens}{number of \code{rmetasim} generations to establish 
linkage disequilibrium}

\item{label}{label for the output folder. If \code{NULL}, the label will be 
"\code{sim.results.YYYYMMDD.HHMM}".}

\item{fsc.exec}{name of fastsimcoal executable}

\item{num.cores}{number of cores to use}
}
\value{
invisibly, the label used to name output files
}
\description{
Simulate populations of SNP data following island or stepping 
  stone models
}
\note{
The following arguments can be specified as single values or vectors:
 \code{num.pops, num.loci, div.time, Ne, Nm, theta, mig.type}. One simulation 
 scenario will be generated for each combination of unique values of these 
 arguments.  
 
 The function requires that the executable for \code{fastsimcoal} be installed 
 in a location where it can be executed from the current working directory.  
 
 The function will create a folder in the current working directory with the 
 name provided by \code{label}. Within this folder there will be an R 
 workspace file ("\code{scenarios.rdata}") containing a data frame with the 
 parameters for each scenario (\code{sc.df}), and one R  
 workspace file for each scenario, named \code{"gtypes.#.rdata"} where "#" 
 is the scenario number. Each scenario file contains a list named 
 \code{sim.data}, which contains a \code{strataG::gtypes} representation 
 of the simulated SNP data for every replicate in that scenario. The list 
 has a "\code{scenario}" attribute which is a one row data.frame providing 
 the parameters that generated the data and a "\code{label}" attribute which 
 is the label of the set of scenarios that were run.
}
\author{
Eric Archer <eric.archer@noaa.gov>
}
