% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grur_imputations.R
\name{split_imp}
\alias{split_imp}
\title{split_imp}
\usage{
split_imp(x, cpu.rounds, parallel.core = parallel::detectCores() - 1)
}
\description{
Split markers into chunk for parallel imputations processing
}
\keyword{internal}
