% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sim.R
\name{tidy_sim}
\alias{tidy_sim}
\title{Tidy simulations}
\usage{
tidy_sim(sim.directory, write.tidy = TRUE,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{sim.directory}{(character, path) Full path to the directory containing simulation objects.}

\item{write.tidy}{(logical, optional) The function uses \code{\link[fst]{write.fst}},
to write the tidy data frame in the appropriate scenario directory.
The file extension is \code{.rad}.
Default: \code{write.tidy = TRUE}.}

\item{parallel.core}{(integer, optional) The number of core used for parallel
execution during extraction of simulated data.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
A list with all the tidy data frames for all scenarios and iterations 
generated by \code{\link{simulate_rad}}. Depending on argument values, the function also 
write the tidy data in the appropriate folder.
}
\description{
Will extract and tidy all simulations objects from a directory,
provided that the simulations were generated with \code{\link{simulate_rad}}.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}
}
