% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{whitelists_markers_generator}
\alias{whitelists_markers_generator}
\title{whitelists_markers_generator}
\usage{
whitelists_markers_generator(x, y, path.folder)
}
\description{
Generate whitelists of markers
}
\keyword{internal}
