% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grur_dapc_pipeline.R
\name{write_dapc_plot}
\alias{write_dapc_plot}
\title{write_dapc_plot}
\usage{
write_dapc_plot(x, plot.name, width = 25, height = 10, dpi = 600)
}
\description{
Write the admixture dapc plots
}
\keyword{internal}
