% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_missing.R
\name{generate_missing}
\alias{generate_missing}
\title{Generate missing data}
\usage{
generate_missing(data, output, filename = NULL, memorize.missing = NULL,
  missing.data.type = "MCAR", prop.missing.overall = 0.3,
  average.read.depth = 10, min.reads = 6, alpha.beta.markers = c(5, 40),
  random.seed = NULL, imputation.method = NULL,
  hierarchical.levels = "populations", verbose = TRUE,
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{data}{11 options: VCF (SNPs or Haplotypes,
to make the vcf population ready, see details below),
plink, stacks haplotype file, genind (library(adegenet)),
genlight (library(adegenet)), gtypes (library(strataG)), genepop,
and a data frame in long/tidy or wide format.
\emph{See details} of \code{\link{tidy_genomic_data}}.}

\item{output}{16 genomic data formats can be exported: tidy (by default),
genind, genlight, vcf (biallelic, for file format version, see details below), plink, genepop,
structure, arlequin, hierfstat, gtypes (strataG), betadiv, snprelate (for SNPRelate's GDS).
Use a character string,
e.g. \code{output = c("genind", "genepop", "structure")}, to have preferred
output formats generated. The tidy format is generated automatically.
Default: \code{output = NULL}.}

\item{filename}{(optional) The filename prefix for the objet in the global environment
or the working directory. Default: \code{filename = NULL}. A default name will be used,
customized with the output file(s) selected.}

\item{memorize.missing}{(optional, string) To use missing pattern
taken from the function \code{\link[grur]{memorize_missing}}. The argument
require 2 values: the path to the file or the name of the object in the
global environment and the column selected to introduce missingness.
e.g. \code{memorize.missing = c("~/missing_analysis/missing.memory", "MISSING_INDIVIDUALS_MIX")}.
See function documentation for more details on columns naming scheme.
Default: \code{memorize.missing = NULL}.}

\item{missing.data.type}{(character) Three options 
(based on Little, R. J., & Rubin, D. B., 2002):
\enumerate{
\item \strong{MCAR} (Missing Completely At Random): the missingness
probability does not depend on any observed or unobserved data
(independent of any covariates).
The risk of introducing biais during imputations is very very low.

\item \strong{MAR} (Missing At Random): the missingness probability
depends only on the observed data.

\item \strong{NMAR} (Not Missing At Random): the missingness probability
depends on both the observed and unobserved data.
}
Default: \code{missing.data.type = "MCAR"}.}

\item{prop.missing.overall}{(double) The overall proportion of missing genotypes
generated.
Default: \code{prop.missing.overall = 0.3}.}

\item{average.read.depth}{(integer) Desired average read depth at a marker in an individual.
Default: \code{average.read.depth = 10}.}

\item{min.reads}{(integer) Minimum number of simulated reads to call a SNP.
Everything below will results in missing genotypes.
Default: \code{min.reads = 6}.}

\item{alpha.beta.markers}{(string of 2 integers) The shape and scale parameter
for the gamma distribution over markers.
Default: \code{alpha.beta.markers = c(5,40)}.}

\item{random.seed}{(integer) Random seed number for reproducibility.
With default: \code{random.seed = NULL} the number is generated automatically and randomly.
for the input file and \code{\link[stackr]{genomic_converter}} for the output file parameter.}

\item{imputation.method}{(character, optional)
Methods available for map-independent imputations of missing genotype:

(1) \code{imputation.method = "max"} Strawman imputation,
the most frequently observed genotypes (ties are broken at random).

(2) \code{imputation.method = "rf"} On-the-fly-imputations using
Random Forests algorithm.

(3) \code{imputation.method = "boost"} extreme gradient boosting trees.

(4) \code{imputation.method = "mca"} Multiple Correspondence Analysis.

\code{imputation.method = NULL} will return the original dataset, without
imputation.
Default: \code{imputation.method = "rf"}.}

\item{hierarchical.levels}{(character, optional) \code{c("global", "populations")}.
Should the imputations be computed globally or by populations. Note that
imputing genotype globally in conjunction with \code{imputation.method = "max"}
can create huge bias for example by
introducing foreign genotypes in some populations (see note for more info).
Default: \code{hierarchical.levels = "populations"}.}

\item{verbose}{(optional, logical) When \code{verbose = TRUE}
the function is a little more chatty during execution.
Default: \code{verbose = TRUE}.}

\item{parallel.core}{(optional) The number of core used for parallel
execution during vcf import.
Default: \code{parallel::detectCores() - 1}.}
}
\value{
In the global environment, a list with the tidy data set, the random.seed and function.call.
In the working directory, the output file with format selected.
}
\description{
Artificially induce missing genotypes by generating
a compound Dirichlet-multinomial distribution.
The input and out files see
\href{https://github.com/thierrygosselin/stackr}{stackr} for more details.
Input files are recognized through \code{\link[stackr]{tidy_genomic_data}}
and output files are generated with and without imputations through
\code{\link[stackr]{genomic_converter}}.
}
\examples{
\dontrun{
The simplest form of the function:

datamissing <- generate_missing(
data = "plink.tped",
output = "genepop"
)
}
}
\references{
Little, R. J., & Rubin, D. B. (2002) Statistical analysis with missing data.
}
\author{
Thierry Gosselin \email{thierrygosselin@icloud.com}, Eric C. Anderson \email{eric.anderson@noaa.gov} and Greg L. Owens \email{gregory.owens@alumni.ubc.ca}
}
