% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_missing.R
\name{reads_per_markers}
\alias{reads_per_markers}
\title{reads_per_markers}
\usage{
reads_per_markers(total.reads, markers.cdm, markers.list)
}
\description{
Function to apply to each ind and markers
}
\keyword{internal}
