% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rad.R
\name{simulate_rad}
\alias{simulate_rad}
\title{Simulate RADseq data}
\usage{
simulate_rad(num.pops = 5, num.loci = 1000, div.time = 25000,
  ne = c(50, 500), nm = c(0, 0.1, 0.5, 1, 5), theta = 0.2,
  mig.type = c("island", "stepping.stone"), num.reps = 3,
  num.rms.gens = 10, label = NULL, fsc.exec = "fsc26",
  parallel.core = parallel::detectCores() - 1)
}
\arguments{
\item{num.pops}{(integer) Number of populations.
Default: \code{num.pops = 5}.}

\item{num.loci}{(integer) Number of independent RADseq loci.
To use more than 2000 loci, \pkg{rmetasim} needs to be modified
(\href{https://github.com/thierrygosselin/grur/blob/master/vignettes/vignette_grur.Rmd}{vignette}).
Default: \code{num.loci = 1000}.}

\item{div.time}{(integer) Time since divergence of populations.
Default: \code{div.time = 25000}.}

\item{ne}{(integer) Effective populations size.
Default: \code{ne = c(50, 500)}.}

\item{nm}{(integer) Number of migrants per generation.
Default: \code{nm = c(0, 0.1, 0.5, 1, 5)}.}

\item{theta}{(integer) Value of theta (= ne * migration rate).
Default: \code{theta = 0.2}.}

\item{mig.type}{(integer) Migration topology type: \code{"island"} or 
\code{"stepping.stone"}.}

\item{num.reps}{(integer) Number of replicates for each scenario.
Default: \code{num.reps = 10}.}

\item{num.rms.gens}{(integer) Number of \code{rmetasim} generations to establish
linkage disequilibrium.
Default: \code{num.rms.gens = 10}.}

\item{label}{(integer) Label for the output folder. With default \code{label = NULL},
the label will be "\code{sim.results.YYYYMMDD.HHMM}".}

\item{fsc.exec}{(character) Name of fastsimcoal executable. See details.
Default: \code{fsc.exec = "fsc26"} (for latest version: v.2.6.0.3 - 14.10.17)}

\item{parallel.core}{(optional, integer) The number of core used for parallel
execution.
Default: \code{parallel.core = parallel::detectCores() - 1}.}
}
\value{
invisibly, the label used to name output files
}
\description{
Simulate populations of RADseq data following island or stepping 
stone models.
Inside the function, allele frequency can be created with
\href{http://cmpg.unibe.ch/software/fastsimcoal2/}{fastsimcoal2} and then used
inside \href{https://github.com/stranda/rmetasim}{rmetasim} simulation engine.
The function requires \href{https://github.com/EricArcher/strataG}{strataG}
to be installed.
}
\details{
\strong{fastsimcoal2: }
 The function requires that the executable for 
 \href{http://cmpg.unibe.ch/software/fastsimcoal2/}{fastsimcoal} be installed 
 in a location where it can be executed from the current working directory.
 
 \href{http://gbs-cloud-tutorial.readthedocs.io/en/latest/03_computer_setup.html?highlight=PATH#save-time}{How to set up your PATH}.
 
 \href{http://gbs-cloud-tutorial.readthedocs.io/en/latest/07_start_from_new_image.html#install-gbs-radseq-software-time-required-30min}{How to install fastsimcoal2 v.2.6.0.3}
}
\note{
The following arguments can be specified as single values or vectors:
 \code{num.pops, num.loci, div.time, ne, nm, theta, mig.type}. One simulation 
 scenario will be generated for each combination of unique values of these 
 arguments.  
 
 
 The function will create a folder in the current working directory with the 
 name provided by \code{label}. Within this folder there will be an R 
 workspace file ("\code{scenarios.rdata}") containing a data frame with the 
 parameters for each scenario (\code{sc.df}), and one R  
 workspace file for each scenario, named \code{"gtypes.#.rdata"} where "#" 
 is the scenario number. Each scenario file contains a list named 
 \code{sim.data}, which contains a \code{strataG::gtypes} representation 
 of the simulated SNP data for every replicate in that scenario. The list 
 has a "\code{scenario}" attribute which is a one row data.frame providing 
 the parameters that generated the data and a "\code{label}" attribute which 
 is the label of the set of scenarios that were run.
}
\examples{
\dontrun{
require(strataG)
require(rmetasim)
library(grur)
sim <- grur::simulate_rad(
num.pops = 3, num.loci = 200, div.time = 1000, ne = c(50,200),
nm = c(0, 0.1), theta = 0.2, mig.type = c("island", "stepping.stone"), 
num.reps = 3, num.rms.gens = 5, label = NULL, fsc.exec = "fsc26", 
parallel.core = 4)
}
}
\seealso{
\href{https://github.com/EricArcher/strataG}{strataG}
}
\author{
Eric Archer \email{eric.archer@noaa.gov} and Thierry Gosselin \email{thierrygosselin@icloud.com}
}
